error fixing
isFIXED => edit product cant edit if same name which defeats the purpose
isDONE => delete search bar on top with notif icon button

isFIXED => edit batch order rename makes it error
isDONE => batch order search bar dont work

isFIXED => edit batch order updating the amount makes it error
isDONE => edit batch order - updated it to received but doesn't reflect on data grid view (pending)
+++ added update PENDING STATUS

isFIXED => delete employee errors
isFIXED => update employee doesn't work?

isFIXED => WIDGET PROBLEM
add product - add button adds to db but causes error causing program to exit
add employee - add button adds to db but causes error

+++ Dashboard_Version1.0/Dashboard_Version1.0/MainDashboard/Frontend/EmployeeStaff/Employee.cs
@@ -56,7 +56,7 @@ namespace MainDashboard
             string staffName = selectedRow.Cells["EmployeeName"].Value?.ToString()?.Trim() ?? "";
 
             var staffReader = new StaffCrud();
-            int? staffId = staffReader.GetStaffIDByName("John Doe");
+            int? staffId = staffReader.GetStaffIDByName(staffName);
 
             if (clickedColumnName == "Edit")
             {
@@ -75,10 +75,20 @@ namespace MainDashboard
 
                 if (confirmResult == DialogResult.Yes)
                 {
+
+                    if (staffId == null)
+                    {
+                        MessageBox.Show("Staff ID not found or is invalid.", "Deletion Error", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        return;
+                    }
+
+                    var authDeleteHandler = new AuthDelete();
+                    bool deletingAuthSuccess = authDeleteHandler.DeleteAuthByStaffID(staffId.Value);
+
                     var deleteHandler = new StaffDelete();
                     bool deletingStaffSuccess = deleteHandler.DeleteStaffByID(staffId.Value);
 
-                    if (deletingStaffSuccess)
+                    if (deletingAuthSuccess && deletingStaffSuccess)
                     {
                         MessageBox.Show("Successfully deleted the employee.", "Deletion Window",
                             MessageBoxButtons.OK, MessageBoxIcon.Information);
